requirements.txt
streamlit==1.38.0
pandas==2.2.2
requests==2.31.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
openpyxl==3.1.2
xlsxwriter==3.2.0
